Revert "Revert "change hexToInt to 32-bit""
This reverts commit b9e700630b26c2f39e406590250cad01b0fa68d0.

+++ chapter_16/Text_Listings/convertHex_C/hexToInt.c
@@ -1,4 +1,4 @@
-// Converts hex character string to long int
+// Converts hex character string to int
 // Returns number of characters
  
 #include "hexToInt.h"
@@ -6,9 +6,9 @@
 #define INTPART 0x0f  // also works for lowercase
 #define NUL '\0'
 
-int hexToInt(long int *intPtr, char *stringPtr) {
+int hexToInt(int *intPtr, char *stringPtr) {
     char current;
-    long int result;
+    int result;
     int count;
 
     count = 0;
